+++ frontend/src/pages/DisplayPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import DisplayTableContainer from "../crud/DisplayTableContainer";
-import MainPageButton from "../common/MainPageButton";
+import DisplayTableContainer from "../components/crud/DisplayTableContainer";
+import MainPageButton from "../components/common/MainPageButton";
 
 const GetPage = (): React.ReactElement => {
   return (
