+++ frontend/src/pages/HooksDemo/EmojiSection.tsx
