+++ frontend/src/pages/ProfilePage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Text } from "@chakra-ui/react";
-import NavBar from "../common/navbar/NavBar";
-import Logout from "../auth/Logout";
+import NavBar from "../components/common/navbar/NavBar";
+import Logout from "../components/auth/Logout";
 
 const ProfilePage = (): React.ReactElement => {
   return (
