+++ frontend/src/pages/SimpleEntityDisplayPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import SimpleEntityDisplayTableContainer from "../crud/SimpleEntityDisplayTableContainer";
-import MainPageButton from "../common/MainPageButton";
+import SimpleEntityDisplayTableContainer from "../components/crud/SimpleEntityDisplayTableContainer";
+import MainPageButton from "../components/common/MainPageButton";
 
 const GetSimpleEntitiesPage = (): React.ReactElement => {
   return (
