+++ frontend/src/pages/SimpleEntityUpdatePage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import SimpleEntityUpdateForm from "../crud/SimpleEntityUpdateForm";
-import MainPageButton from "../common/MainPageButton";
+import SimpleEntityUpdateForm from "../components/crud/SimpleEntityUpdateForm";
+import MainPageButton from "../components/common/MainPageButton";
 
 const SimpleEntityUpdatePage = (): React.ReactElement => {
   return (
