+++ frontend/src/pages/UserManagementPage.tsx
@@ -13,10 +13,10 @@ import {
   AlertIcon,
   CloseButton,
 } from "@chakra-ui/react";
-import UserAPIClient from "../../APIClients/UserAPIClient";
-import { User } from "../../types/UserTypes";
-import MainPageButton from "../common/MainPageButton";
-import AddUserFormModal, { AddUserRequest } from "../crud/AddUserFormModal";
+import UserAPIClient from "../APIClients/UserAPIClient";
+import { User } from "../types/UserTypes";
+import MainPageButton from "../components/common/MainPageButton";
+import AddUserFormModal, { AddUserRequest } from "../components/crud/AddUserFormModal";
 
 const handleUserSubmit = async (formData: AddUserRequest) => {
   // eslint-disable-next-line no-useless-catch
